Refactors based on PR comments
