+++ src/SearchQueryService/Indexes/Models/Solr/ISolrField.cs
@@ -1,0 +1,4 @@
+﻿namespace SearchQueryService.Indexes.Models
+{
+    public interface ISolrField { }
+}
