+++ src/SearchQueryService/Indexes/Models/Solr/ResponseHeader.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace SearchQueryService.Indexes.Models
+{
+    public class ResponseHeader
+    {
+        public int Status { get; set; }
+        public int QTime { get; set; }
+    }
+}
