+++ src/SearchQueryService/Indexes/Models/Solr/SearchIndex.cs
@@ -1,0 +1,8 @@
+﻿namespace SearchQueryService.Indexes.Models
+{
+    public class SearchIndex
+    {
+        public string Name { get; set; }
+        public AzField[] Fields { get; set; }
+    }
+}
