+++ src/SearchQueryService/Indexes/Models/Solr/SearchResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace SearchQueryService.Indexes.Models
+{
+    public class SearchResponse
+    {
+        public ResponseHeader ResponseHeader { get; set; }
+        public Response Response { get; set; }
+    }
+}
